fix: Implement full emoji support

+++ internal/texture.hpp
@@ -20,13 +20,13 @@ namespace cppreact {
 
   /** @brief Abstract base class for texture resources. */
   class texture {
-    int _width;  ///< Texture width in pixels
-    int _height; ///< Texture height in pixels
+    uint16_t _width;  ///< Texture width in pixels
+    uint16_t _height; ///< Texture height in pixels
     public:
     texture(int width, int height) : _width(width), _height(height) {}
     virtual ~texture() {}
-    int width() {return _width;}
-    int height() {return _height;}
+    uint16_t width() {return _width;}
+    uint16_t height() {return _height;}
     /** @brief Get the color of a single pixel. */
     virtual color get_pixel(int x, int y) = 0;
     /** @brief Get raw pixel data for the entire texture. */
